% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{calendrier_climatique_custom}
\alias{calendrier_climatique_custom}
\title{Calendrier climatique journalier custom}
\usage{
calendrier_climatique_custom(
  data,
  weather_data_j,
  id_col,
  date_semi = date_semi,
  date_recolte = date_recolte,
  nb_jour
)
}
\arguments{
\item{data}{Données de rendement contenant au moins une colonne d’identification des parcelles, une date de semi, un numéro de champ et une année. L'année peut être extraite de la date de semi.}

\item{weather_data_j}{Données climatiques journalières contenant au moins la température minimale et maximale et les précipitations totales.}

\item{id_col}{Variable d'identification des parcelles/unité de culture}

\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}.}

\item{date_recolte}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}.}

\item{nb_jour}{Nombre de jours depuis la date de semis.}
}
\value{
Un \code{dataframe} contenant le calendrier climatique de date -> date + nb_jour pour chaque date de semis.
}
\description{
Transforme un ensemble de dates de semis et de récolte en un calendrier climatique pour l'intervalle de chaque combinaison de dates.
Chaque combinaison de dates est étendue par incrément d'une journée où les données climatiques journalières y sont jointes, et ce, pour trois scénarios distincts.
}
