% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{calendrier_climatique_horaire}
\alias{calendrier_climatique_horaire}
\title{Calendrier climatique horaire}
\usage{
calendrier_climatique_horaire(
  data,
  weather_data_h,
  id_col,
  date_semi = date_semi
)
}
\arguments{
\item{data}{Données de rendement contenant au moins une colonne d'identification des parcelles, une date de semi, un numéro de champ et une année. L'année peut être extraite de la date de semi.}

\item{weather_data_h}{Données climatiques horaires contenant au moins la température}

\item{id_col}{Variable d'identification des parcelles/unité de culture}

\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}.}
}
\value{
Un \code{dataframe} contenant le calendrier climatique de date -> date + 14 pour chaque date de semis.
}
\description{
Transforme un ensemble de dates de semis en un calendrier climatique.
Chaque date de semis est étendue par incrément d'une heure jusqu'à \code{date_semi + 14} où les données climatiques horaires y sont jointes.
Ce calendrier est utilisé pour calculer les périodes de gel.
}
