% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{frequence_gel}
\alias{frequence_gel}
\title{Fréquence des épisodes de gel}
\usage{
frequence_gel(calendrier, id_col, date_semi = date_semi, thresh)
}
\arguments{
\item{calendrier}{Calendrier climatique journalier obtenu avec la fonction \code{calendrier_climatique_journalier}. La période d'intérêt est données par \code{Postsemis_14j}.}

\item{id_col}{Variable d'identification des parcelles/unité de culture}

\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}

\item{thresh}{Température minimale de résistance au gel de la culture, variable par culture}
}
\value{
Un \code{dataframe} contenant:
\describe{
\item{id_col}{Variable d'identification des parcelles/unité de culture}
\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}
\item{nochamp}{Variable d'identification des champs}
\item{annee}{Variable d'identification de l'année}
\item{cold}{Nombre d'épisodes de gel entre \code{date_semi} et \code{date_semi + 14}. Retourne 0 lorsque qu'aucun épisode de gel n'a été enregistré.}
}
}
\description{
Cette fonction dénombre le nombre d'épisodes de gel pour chaque date de semis entre \code{date_semi} et \code{date_semi + 14}.
Un épisode de gel est caractérisé par une température minimale journalière sous un seuil (\code{thresh}).
}
