% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{gdd}
\alias{gdd}
\title{Calcul des degrés jour}
\usage{
gdd(min_temp, max_temp, tbase = 5, tlim = 32)
}
\arguments{
\item{min_temp}{Température minimale journalière}

\item{max_temp}{Température maximale journalière}

\item{tbase}{Température basale où l'accumulation des degrés jour commence, variable par culture}

\item{tlim}{Température maximale où l'accumulation des degrés jour arrête, variable par culture}
}
\description{
Calcul des degrés jour
}
