% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{gdd_cummul}
\alias{gdd_cummul}
\title{Cummul des degrés jour}
\usage{
gdd_cummul(calendrier, id_col, date_semi, min_temp, max_temp, tbase, tlim)
}
\arguments{
\item{calendrier}{Calendrier climatique journalier du semis à la récolte}

\item{id_col}{Variable d'identification des parcelles/unité de culture}

\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}

\item{min_temp}{Variable contenant la température minimale journalière}

\item{max_temp}{Variable contenant la température maximale journalière}

\item{tbase}{Température basale où l'accumulation des degrés jour commence, variable par culture}

\item{tlim}{Température maximale où l'accumulation des degrés jour arrête, variable par culture}
}
\value{
Un \code{dataframe} contenant:
\describe{
\item{id_col}{Variable d'identification des parcelles/unité de culture}
\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}
\item{nochamp}{Variable d'identification des champs}
\item{annee}{Variable d'identification de l'année}
\item{unite_chaleur_atteint}{Nombre de degrés jours atteints à la récolte.}
}
}
\description{
Les degrés jour sont utilisés comme barème d'atteinte de maturité physiologique chez le blé.
Cette fonction retourne le nombre de degrés jour atteints au nombre de jours requis pour l'atteinte de la maturité physiologique.
}
