% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{precipitation}
\alias{precipitation}
\title{Cummul des précipitation}
\usage{
precipitation(calendrier, id_col, date_semi = date_semi)
}
\arguments{
\item{calendrier}{Liste de calendriers climatiques journaliers. Cette fonction accepte le output de la fonction \code{calendrier_climatique_journalier}}

\item{id_col}{Variable d'identification des parcelles/unité de culture}

\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}
}
\value{
Une liste de \code{dataframe} contenant:
\describe{
\item{Presemis_7j}{Un calendrier contenant les précipitations cummulées 7 jours avant le semis}
\item{Postsemis_14j}{Un calendrier contenant les précipitations cummulées 14 jours après le semis}
\item{Recolte}{Un calendrier contenant les précipitations cummulées du semis à la récolte moins un jour}
}
}
\description{
Cette fonction cumule les précipitations selon trois scénarios distincts. Une variable contenant les précipitations totales journalière \code{total_precip} doit absolument être incluse.
}
