% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{rad_cummul}
\alias{rad_cummul}
\title{Cummul de la radiation solaire}
\usage{
rad_cummul(calendrier, id_col, date_semi)
}
\arguments{
\item{calendrier}{Calendrier climatique journalier du semis à la récolte}

\item{id_col}{Variable d'identification des parcelles/unité de culture}

\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}
}
\value{
Un \code{dataframe} contenant:
\describe{
\item{id_cols}{Variable d'identification des parcelles/unité de culture}
\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}
\item{nochamp}{Variable d'identification des champs}
\item{annee}{Variable d'identification de l'année}
\item{globalrad}{Somme de la radiation globale, unités variables selon la source des données.}
}
}
\description{
Cummul de la radiation solaire
}
