% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{utm}
\alias{utm}
\title{Calcul des UTMs}
\usage{
utm(min_temp, max_temp)
}
\arguments{
\item{min_temp}{Température minimale journalière}

\item{max_temp}{Température maximale journalière}
}
\description{
Les unités thermiques maïs sont utilisées comme barème d'atteinte de maturité physiologique chez le soya et le maïs grain.
}
