% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_meteo.R
\name{utm_cummul}
\alias{utm_cummul}
\title{Cummul des UTM}
\usage{
utm_cummul(calendrier, id_col, date_semi, min_temp, max_temp)
}
\arguments{
\item{calendrier}{Calendrier climatique journalier du semis à la récolte}

\item{id_col}{Variable d'identification des parcelles/unité de culture}

\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}

\item{min_temp}{Variable contenant la température minimale journalière}

\item{max_temp}{Variable contenant la température maximale journalière}
}
\value{
Un \code{dataframe} contenant:
\describe{
\item{id_cols}{Variable d'identification des parcelles/unité de culture}
\item{date_semi}{Variable de type \code{date} sous format \verb{"\%Y-\%m-\%d}}
\item{nochamp}{Variable d'identification des champs}
\item{annee}{Variable d'identification de l'année}
\item{unite_chaleur_atteint}{Nombre d'UTM atteints à la récolte.}
}
}
\description{
Les unités thermiques maïs sont utilisées comme barème d'atteinte de maturité physiologique chez le soya et le maïs grain.
}
